Implemented initial thread burst in stepping thread group.
This allows to quickly reach a target number of threads to
then iterate with finer steps around that number.

The ramp up used for the initial burst is the same as for
the other steps.

The plugin is backward compatible and the initial thread burst
defaults to 0, which replicates the old behavior.