Changed JsonPathExtractor to extract single results without array brackets.
When a JSON path filter is applied on an array, the result is always another array. This is undesirable when only a single value is matched because additional processing is needed to remove the brackets around the result before it can be inserted into a subsequent request.

Instead of returning ["value"], only "value" will now be extracted. If more than one element matches (i.e. [1,"two",3]) the brackets
are retained.

Added unit tests for the new behavior.
Refactored existing unit tests, added actual assertions to all test cases.