Added capability to set Y-Axis label on perfmon charts.
Having possibility to label Y-Axis in performance tests tool is a must. Otherwise each time report is sent, people are asking what are units on Y axis? Its According to my own experience with perf tests and this:
https://groups.google.com/forum/#!topic/jmeter-plugins/zjTwY38X_z8

I made this change way to not disrupt other things, so if no --yAxisLabel is specified, it uses default value, which is "Performance Metrics" like now, otherwise this label is set.