RTE-Plugin v3.1
What's new on this minor release:

•Fix issue when sending Attention Key on non-fields screen.

•Add support for sending an attention key at a determined position. This feature is very useful for scenarios where screens aren't constituted by fields, and there is a menu waiting for an ENTER key to be pressed on the desired position.
In order to set the cursor position when building manually the TestPlan, would be adding a Position Input with the coordinates where we want to place the cursor, and then, is mandatory to set the value of the input, empty. The Recorder will do this automatically. For more information click here

•Fix issue when the sample name kept changing when sending inputs using VT420 protocol alongside with the Recorder.