Remove public instances of org.apache.log.Logger
As JMeter has dropped the dependence on the org.apache.log.Logger class, we should
remove it from jmeter-plugins, too. As the change from the old logkit style Logger
to something different will break serializability/compatibility anyway, we can
change the logger to a private member of the class, while we are at it. This
will - hopefully - make it easier in the future to change the logging framework,
if needed.