Remove old LoggingManager and Logger (#426)
* Change private logger to SLF4J

The class org.apache.jorphan.logging.LoggingManager has been deprecated in Apache JMeter
and will be removed. As the default logging interface in JMeter is SLF4J, this patch
changes all private loggers to the same interface.

The conversion was done by the following script:

  find . -name "*.java" -type f \
   | xargs -l1 grep -l "private static final Logger" \
   | while read i; do
       sed -i 's/private static final Logger \(LOGGER\|log\) = LoggingManager.getLoggerForClass();/private static final Logger \1 = LoggerFactory.getLogger('$(basename $i .java)'.class);/; s/org.apache.jorphan.logging.LoggingManager/org.slf4j.LoggerFactory/; s/org.apache.log.Logger/org.slf4j.Logger/;' $i;
     done

To make the code compile, the few occurrences LoggingManager#setPriority have been removed.

* Remove public instances of org.apache.log.Logger

As JMeter has dropped the dependence on the org.apache.log.Logger class, we should
remove it from jmeter-plugins, too. As the change from the old logkit style Logger
to something different will break serializability/compatibility anyway, we can
change the logger to a private member of the class, while we are at it. This
will - hopefully - make it easier in the future to change the logging framework,
if needed.